refactor(core): extract shared FastConformer encoder core, switch parakeet-ctc
parakeet-ctc and parakeet-tdt built byte-for-byte identical dw-striding
subsampling + conformer-stack encoders (arena alloc/upload/bind, BatchNorm
fold, zero-bias synthesis, f16 depthwise upload), differing only in their
tail (CTC head vs. joint encoder projection) and a couple of checkpoint
metadata knobs (bias presence, scale_input). Extract that skeleton into
models::fastconformer (graph.rs + weights.rs), generic over each family's
own error enum via two small marker traits, with the tail threaded through
as declare/upload closures so it participates in the arena's single
"declare everything, then upload everything" pass.

Switch parakeet-ctc onto the shared core; parakeet-tdt still carries its
own (now-duplicate) copy and moves in the next commit. Numeric behavior is
carried over verbatim -- only where the code lives changes.
