+++ crates/openasr-core/src/models.rs
@@ -10,6 +10,7 @@ pub(crate) mod decode_token_history;
 pub(crate) mod diarize_pack_import;
 pub(crate) mod dolphin;
 pub(crate) mod executor_component_registry;
+pub(crate) mod fastconformer;
 pub(crate) mod firered_aed;
 pub(crate) mod firered_punc;
 pub(crate) mod frame_sync_streaming_driver;
